Change content[:record_id] to :id and update moves endpoint
Record blocks have had their content changed from :record_id to just :id

The moves endpoint no longer just takes a position parameter, it takes
the id of the block you wish to move the item you're moving to. This is
because the frontend only has access to those two ids and not the new
position when it performs the API request.