feat(fix for integer facets): Update the way that we request integer facets
Changed SJ to index a integer and a string field for fields that are
integers that are set as [:filters] in the schema.
This is because we have an issue with docValues where the IntPointField
cannot be used for facetting without having docValues built, but
docValues are providing an inaccurate facet count result.

Updated the exclude_filters_from_facet flag to only add filters as
facets that are considered valid facets and also that have been
requested in the facet block.

references (Optional)