fix(Facet exclusions): correctly account for nested OR queries
Follow-up to https://github.com/DigitalNZ/supplejack_api/pull/232.

`exclude_filters_from_facets` was not working correctly with nested OR queries, because the `elsif` condition meant to catch them was incorrect.
In fact, a nested OR query is not a `Hash`, but rather a `ActiveSupport::HashWithIndifferentAccess`.
Therefore it was failing the class check (but not in the test suite, since in there they are regular hashes).

Now, the `elsif` condition works correctly :)