fix(StoryItemSerializer): returning deleted record status
This is interesting. Hear me out!

record_id for the record of the set_item is placed in set_item[:content][:id].
Early on this was in set_item[:content][:record_id]. So some old set items have it wrong.
When the serializer tried to fetch the id for these old items it got nil for record_id.
When nil was used to fetch the record SupplejackApi::Record.find_by(record_id: record_id).
Here comes the kicker, it found a record in DNZ api
<SupplejackApi::Record _id: 51f41917147d7b090110d294, status: "deleted", record_type: 1, record_id: nil>

Hence this record was returned as a set item breaking the front end as this is a deleted record.