fix(IndexProcessor): Race condition
If a second can be broken down to 3 parts and if the following happens in those 3 parts sequentially.
- Enrichment 1 updates record x
- IndexProcessor pics up the record x
- Enrichment 2 updates record x
The second update wont be captured on indexing as the updated_at for both enrichments are at the same second.
So if we query ready_for_indexing with a delay of 5 seconds we make sure that both updates written in any second is captured.